chore(devenv): introduce theme editor
If you click on the icon at the top right, a side bar comes up where
you can change the theme (the capability the earlier PR introduced) as
well as edit the theme. If you go to the bottom in the side bar, you’ll
see a button where the theme content is copied to clipboard.